Uploading file using multer, creating PDF using pdfkit and downloading files

+++ createServer/controllers/admin.js
@@ -1,7 +1,8 @@
 const { validationResult } = require('express-validator');
+const { trusted } = require('mongoose');
 
 const Product = require('../models/product');
-const { trusted } = require('mongoose');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
     res.render('admin/edit-product', 
@@ -10,16 +11,30 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+
+    if(!image){
+        return res.status(422).render('admin/edit-product', 
+        {   product: {
+                title,
+                price, 
+                description
+            }, 
+            pageTitle: 'Add Product', 
+            editing: false,
+            hasError: true,
+            errorMessage: 'Attached file is not an image.',
+            validationErrors: []
+        });
+    }
 
     const errors = validationResult(req);
     if(!errors.isEmpty()){
         return res.status(422).render('admin/edit-product', 
         {   product: {
                 title,
-                imageUrl, 
                 price, 
                 description
             }, 
@@ -30,6 +45,8 @@ exports.postAddProduct = (req, res, next) => {
             validationErrors: errors.array()
         });
     }
+
+    const imageUrl = image.path;
 
     const product = new Product({ userId: req.user, title, price, imageUrl, description });
 
@@ -68,7 +85,7 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
     const productID = req.body.productId;
     const updatedTitle = req.body.title;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedPrice = req.body.price;
     const updatedDescription = req.body.description;
 
@@ -77,7 +94,6 @@ exports.postEditProduct = (req, res, next) => {
         return res.status(422).render('admin/edit-product', 
         {   product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl, 
                 price: updatedPrice, 
                 description: updatedDescription,
                 _id: productID
@@ -97,7 +113,10 @@ exports.postEditProduct = (req, res, next) => {
             }
             product.title = updatedTitle;
             product.price = updatedPrice;
-            product.imageUrl = updatedImageUrl;
+            if(image){
+                fileHelper(product.imageUrl);
+                product.imageUrl = image.path;
+            }
             product.description = updatedDescription;
 
             return product.save()
@@ -114,15 +133,24 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const productID = req.body.productId;
-    Product.deleteOne({_id: productID, userId: req.user._id})
-        .then(() => {
-            res.redirect('/admin/products');
-        })
-        .catch(err=>{
-            const error = new Error(err);
-            error.httpStatusCode = 500;
-            next(error);
-        });
+    Product.findById(productID)
+    .then(product=>{
+        if(!product){
+            return next(new Error('Product not found.'));
+        }
+
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({_id: productID, userId: req.user._id});
+    })
+    .then(() => {
+        res.redirect('/admin/products');
+    })
+    .catch(err=>{
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        next(error);
+    });
+        
 };
 
 exports.getAdminProducts = (req, res, next) => {
